+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## vNEXT (not yet released)
 
+### `@liveblocks/client`
+
+- Fix `ToJson` type losing specific value types for `Record<string, T>` fields
+  in Storage
+
 ## v3.18.1
 
 ### `@liveblocks/react-ui`